sdfgsd

+++ example-app/app/Services/ImageService.php
@@ -2,6 +2,7 @@
 
 namespace App\Services;
 
+use Exception;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Storage;
 
@@ -21,18 +22,26 @@ class ImageService
 
   static function add($table, $image)
   {
+    if(empty($image)) {
+      throw new Exception('Не выбран файл. Выберите файл');
+    }
+
     DB::table($table)->insert(
-        ['path' => $image->store('uploads')]
+      ['path' => $image->store('uploads')]
     );
   }
 
   static function update($table, $imageform, $id)
   {
+    if(empty($imageform))
+    {
+      throw new Exception('Не выбран файл. Выберите выберите файл');
+    }
     // получаем объект из БД
     $image = ImageService::one('images', $id);
     // удаляем изображение из папки по названию
     Storage::delete($image->path);
-
+    
     //загружаем изображение в папку uploads
     $path = $imageform->store('uploads');
 
@@ -52,4 +61,5 @@ class ImageService
     // удаляем строку из БД
     DB::table($table)->where('id', $id)->delete();
   }
+
 }
